requirements.txt
advertorch==0.2.3
numpy==1.23.5
opencv_python==4.6.0.66
Pillow==9.3.0
torch==1.12.0+cu113
torchmetrics==0.10.3
torchvision==0.13.0+cu113
tqdm==4.51.0
